Began trying to implement flight path guidelines

+++ src/main/Coordinates.java
@@ -12,7 +12,7 @@ public class Coordinates {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
@@ -20,7 +20,7 @@ public class Coordinates {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 }
